Refactorización del AdminController
Refactorización del AdminController
