+++ app/Resources/views/admin/miembros/list.html.twig
@@ -1,10 +1,10 @@
 {% extends 'admin/base_list.html.twig' %}
 {% block list_title %}
-    Usuarios Directores
+    Usuarios Miembros
 {% endblock %}
 
 {% block boton_nuevo %}
-    <a href="{{ path('nuevo miembro') }}"><button class="btn btn-info" type="button" style="margin:15px;"><i class="fa fa-plus"></i> Nuevo Director</button></a>
+    <a href="{{ path('nuevo miembro') }}"><button class="btn btn-info" type="button" style="margin:15px;"><i class="fa fa-plus"></i> Nuevo Miembro</button></a>
 {% endblock %}    
 
 {% block list_header %}
