+++ app/config/security.yml
@@ -5,7 +5,10 @@ security:
         FOS\UserBundle\Model\UserInterface: bcrypt
 
     role_hierarchy:
-        ROLE_ADMIN:       ROLE_USER
+        ROLE_MIEMBRO:     ROLE_USER
+        ROLE_DIRECTOR:    ROLE_MIEMBRO
+        ROLE_SOSTENEDOR:  ROLE_DIRECTOR
+        ROLE_ADMIN:       ROLE_SOSTENEDOR
         ROLE_SUPER_ADMIN: ROLE_ADMIN
 
     providers:
